backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python>=4.10.0
torch>=2.6.0
torchvision>=0.21.0
ultralytics>=8.0.206
Pillow>=10.0.1
numpy>=1.24.3
pydantic>=2.8.0
aiofiles==23.2.1
